refactor: split camera utils
